--- packages/seacas/libraries/ioss/src/exo_fac/CMakeLists.txt
@@ -1,21 +1,0 @@
-
-SET(HEADERS "")
-SET(SOURCES "")
-
-SET_AND_INC_DIRS(DIR ${CMAKE_CURRENT_SOURCE_DIR})
-APPEND_GLOB(HEADERS ${DIR}/*.h)
-APPEND_GLOB(SOURCES ${DIR}/*.C)
-
-INCLUDE_DIRECTORIES(
-  "${CMAKE_CURRENT_SOURCE_DIR}/../"
-  "${CMAKE_CURRENT_BINARY_DIR}/../"
-)
-
-SET(IOSS_DEPENDENCIES Ioex)
-
-TRIBITS_ADD_LIBRARY(
-	Ioexo_fac
-	HEADERS	${HEADERS}
-	SOURCES ${SOURCES}
-	DEPLIBS ${IOSS_DEPENDENCIES}
-)
