+++ packages/seacas/libraries/ioss/src/init/CMakeLists.txt
@@ -21,7 +21,6 @@ ENDIF()
 
 IF (${PACKAGE_NAME}_ENABLE_SEACASExodus)
   SET(EXODUSLIB Ioex)
-  SET(EXODUSFAC Ioexo_fac)
 ENDIF()
 
 IF (TPL_ENABLE_CGNS)
@@ -41,7 +40,7 @@ TRIBITS_ADD_LIBRARY(
 	HEADERS	${HEADERS}
 	SOURCES ${SOURCES}
 	DEPLIBS Ioss ${EXODUSLIB}
-	        Iogn Iogs Iohb Iotr ${IOVISLIB} ${EXODUSFAC}
+	        Iogn Iogs Iohb Iotr ${IOVISLIB} 
             ${PAMGENLIB} ${CGNSLIB} ${VISLIB} ${DATAWAREHOUSELIB} ${ADIOS2LIB}
 )
